refactor(invoke): Move notification tasks to a dedicated file
