+++ .gitlab/notify/notify.yml
@@ -32,9 +32,9 @@ notify:
       # The triggering repo should already have its own notification system
       if [ "$CI_PIPELINE_SOURCE" != "pipeline" ]; then
         if [ "$DEPLOY_AGENT" = "true" ]; then
-          invoke -e pipeline.notify --notification-type "deploy"
+          invoke -e notify.send-message --notification-type "deploy"
         else
-          invoke -e pipeline.notify --notification-type "merge"
+          invoke -e notify.send-message --notification-type "merge"
         fi
       else
         echo "This pipeline was triggered by another repository, skipping notification."
@@ -53,4 +53,4 @@ send_pipeline_stats:
     - export GITLAB_TOKEN=$(aws ssm get-parameter --region us-east-1 --name ci.datadog-agent.gitlab_read_api_token --with-decryption --query "Parameter.Value" --out text)
     - export DD_API_KEY=$(aws ssm get-parameter --region us-east-1 --name ci.datadog-agent.datadog_api_key_org2 --with-decryption --query "Parameter.Value" --out text)
     - !reference [.setup_python_mirror_linux]
-    - invoke -e pipeline.send-stats
+    - invoke -e notify.send-stats
