+++ tasks/notify.py
@@ -1,0 +1,200 @@
+import io
+import os
+import traceback
+from collections import defaultdict
+from datetime import datetime
+from typing import Dict
+
+from invoke import task
+from invoke.exceptions import Exit
+
+from tasks.libs.datadog_api import create_count, send_metrics
+from tasks.libs.pipeline_data import get_failed_jobs
+from tasks.libs.pipeline_notifications import (
+    GITHUB_SLACK_MAP,
+    base_message,
+    check_for_missing_owners_slack_and_jira,
+    find_job_owners,
+    get_failed_tests,
+    send_slack_message,
+)
+from tasks.libs.pipeline_stats import get_failed_jobs_stats
+from tasks.libs.types import FailedJobs, SlackMessage, TeamMessage
+
+UNKNOWN_OWNER_TEMPLATE = """The owner `{owner}` is not mapped to any slack channel.
+Please check for typos in the JOBOWNERS file and/or add them to the Github <-> Slack map.
+"""
+
+
+@task
+def check_teams(_):
+    if check_for_missing_owners_slack_and_jira():
+        print(
+            "Error: Some teams in CODEOWNERS don't have their slack notification channel or jira specified!\n"
+            "Please specify one in the GITHUB_SLACK_MAP or GITHUB_JIRA_MAP map in tasks/libs/pipeline_notifications.py."
+        )
+        raise Exit(code=1)
+    else:
+        print("All CODEOWNERS teams have their slack notification channel and jira project specified !!")
+
+
+@task
+def send_message(_, notification_type="merge", print_to_stdout=False):
+    """
+    Send notifications for the current pipeline. CI-only task.
+    Use the --print-to-stdout option to test this locally, without sending
+    real slack messages.
+    """
+    project_name = "DataDog/datadog-agent"
+
+    try:
+        failed_jobs = get_failed_jobs(project_name, os.getenv("CI_PIPELINE_ID"))
+        messages_to_send = generate_failure_messages(project_name, failed_jobs)
+    except Exception as e:
+        buffer = io.StringIO()
+        print(base_message("datadog-agent", "is in an unknown state"), file=buffer)
+        print("Found exception when generating notification:", file=buffer)
+        traceback.print_exc(limit=-1, file=buffer)
+        print("See the notify job log for the full exception traceback.", file=buffer)
+
+        messages_to_send = {
+            "@DataDog/agent-all": SlackMessage(base=buffer.getvalue()),
+        }
+        # Print traceback on job log
+        print(e)
+        traceback.print_exc()
+        raise Exit(code=1)
+
+    # From the job failures, set whether the pipeline succeeded or failed and craft the
+    # base message that will be sent.
+    if failed_jobs.all_mandatory_failures():  # At least one mandatory job failed
+        header_icon = ":host-red:"
+        state = "failed"
+        coda = "If there is something wrong with the notification please contact #agent-platform"
+    else:
+        header_icon = ":host-green:"
+        state = "succeeded"
+        coda = ""
+
+    header = ""
+    if notification_type == "merge":
+        header = f"{header_icon} :merged: datadog-agent merge"
+    elif notification_type == "deploy":
+        header = f"{header_icon} :rocket: datadog-agent deploy"
+    base = base_message(header, state)
+
+    # Send messages
+    for owner, message in messages_to_send.items():
+        channel = GITHUB_SLACK_MAP.get(owner.lower(), None)
+        message.base_message = base
+        if channel is None:
+            channel = "#datadog-agent-pipelines"
+            message.base_message += UNKNOWN_OWNER_TEMPLATE.format(owner=owner)
+        message.coda = coda
+        if print_to_stdout:
+            print(f"Would send to {channel}:\n{str(message)}")
+        else:
+            send_slack_message(channel, str(message))  # TODO: use channel variable
+
+
+@task
+def send_stats(_, print_to_stdout=False):
+    """
+    Send statistics to Datadog for the current pipeline. CI-only task.
+    Use the --print-to-stdout option to test this locally, without sending
+    data points to Datadog.
+    """
+    project_name = "DataDog/datadog-agent"
+
+    try:
+        global_failure_reason, job_failure_stats = get_failed_jobs_stats(project_name, os.getenv("CI_PIPELINE_ID"))
+    except Exception as e:
+        print("Found exception when generating statistics:")
+        print(e)
+        traceback.print_exc(limit=-1)
+        raise Exit(code=1)
+
+    if not (print_to_stdout or os.environ.get("DD_API_KEY")):
+        print("DD_API_KEY environment variable not set, cannot send pipeline metrics to the backend")
+        raise Exit(code=1)
+
+    timestamp = int(datetime.now().timestamp())
+    series = []
+
+    for failure_tags, count in job_failure_stats.items():
+        # This allows getting stats on the number of jobs that fail due to infrastructure
+        # issues vs. other failures, and have a per-pipeline ratio of infrastructure failures.
+        series.append(
+            create_count(
+                metric_name="datadog.ci.job_failures",
+                timestamp=timestamp,
+                value=count,
+                tags=list(failure_tags)
+                + [
+                    "repository:datadog-agent",
+                    f"git_ref:{os.getenv('CI_COMMIT_REF_NAME')}",
+                ],
+            )
+        )
+
+    if job_failure_stats:  # At least one job failed
+        pipeline_state = "failed"
+    else:
+        pipeline_state = "succeeded"
+
+    pipeline_tags = [
+        "repository:datadog-agent",
+        f"git_ref:{os.getenv('CI_COMMIT_REF_NAME')}",
+        f"status:{pipeline_state}",
+    ]
+    if global_failure_reason:  # Only set the reason if the pipeline fails
+        pipeline_tags.append(f"reason:{global_failure_reason}")
+
+    series.append(
+        create_count(
+            metric_name="datadog.ci.pipelines",
+            timestamp=timestamp,
+            value=1,
+            tags=pipeline_tags,
+        )
+    )
+
+    if not print_to_stdout:
+        response = send_metrics(series)
+        if response["errors"]:
+            print(f"Error(s) while sending pipeline metrics to the Datadog backend: {response['errors']}")
+            raise Exit(code=1)
+        print(f"Sent pipeline metrics: {series}")
+    else:
+        print(f"Would send: {series}")
+
+
+# Tasks to trigger pipeline notifications
+
+
+def generate_failure_messages(project_name: str, failed_jobs: FailedJobs) -> Dict[str, SlackMessage]:
+    all_teams = "@DataDog/agent-all"
+
+    # Generate messages for each team
+    messages_to_send = defaultdict(TeamMessage)
+    messages_to_send[all_teams] = SlackMessage(jobs=failed_jobs)
+
+    failed_job_owners = find_job_owners(failed_jobs)
+    for owner, jobs in failed_job_owners.items():
+        if owner == "@DataDog/multiple":
+            for job in jobs.all_non_infra_failures():
+                for test in get_failed_tests(project_name, job):
+                    messages_to_send[all_teams].add_test_failure(test, job)
+                    for owner in test.owners:
+                        messages_to_send[owner].add_test_failure(test, job)
+        elif owner == "@DataDog/do-not-notify":
+            # Jobs owned by @DataDog/do-not-notify do not send team messages
+            pass
+        elif owner == all_teams:
+            # Jobs owned by @DataDog/agent-all will already be in the global
+            # message, do not overwrite the failed jobs list
+            pass
+        else:
+            messages_to_send[owner].failed_jobs = jobs
+
+    return messages_to_send
